refactor: Product 패키징 작업
