+++ src/main/java/kitchenpos/dto/menu/MenuResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.dto.menu;
 
 import kitchenpos.domain.Price;
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.dto.product.ProductResponse;
 
 import static java.util.stream.Collectors.*;
